test: adding test for the signup frontend components -- changed from jest to mocha
