+++ test/unit/register/hybee-signup-test-2/package.json
@@ -2,14 +2,15 @@
   "name": "functionality-test",
   "version": "1.0.0",
   "description": "",
-  "main": "index.js",
+  "main": "util.js",
   "scripts": {
-    "test": "jest"
+    "test": "mocha"
   },
   "keywords": [],
   "author": "",
   "license": "ISC",
   "devDependencies": {
-    "jest": "^26.0.1"
+    "chai": "^4.2.0",
+    "mocha": "^7.2.0"
   }
 }
